Reorder Sets
Progress PR for Issue #41. Currently you can edit the order (long press to reorder) in the workout plan screen. The order change is only local; it persists only for the current session. However, I have not figured out how to reflect the changed order in the DB using the API. Does Day even store the order of its Sets in the DB? Or is it dynamically populated into the model in the app?