await for all setExercisesAndUnits calls
Not setting the "await" was causing the method to be run asynchronously which means
that sometimes these would get called several times for the same exercise, which in
turn would cause exercises to be written several times to the local db, polluting the
exercise list, etc. etc.

We increase the schemaVersion of the database because this causes all the cache tables
to be dropped and re-created